Move wallet identity logic
