+++ rust/src/manager/cloud_backup_manager/actors/restore.rs
@@ -262,7 +262,7 @@ impl RestoreOperation {
             labels_failed_errors: Vec::new(),
         };
 
-        let existing_identities = crate::backup::import::collect_existing_wallet_identities()
+        let existing_identities = crate::wallet_identity::collect_existing_wallet_identities()
             .map_err_prefix("collect wallet identities", CloudBackupError::Internal)?;
 
         let mut restore_session = WalletRestoreSession::new(existing_identities);
