+++ rust/src/manager/cloud_backup_manager/ops/enable.rs
@@ -298,7 +298,7 @@ impl RustCloudBackupManager {
         cloud: &CloudStorageClient,
         namespaces: &[MergeNamespace],
     ) -> Result<Vec<MergedNamespaceWallets>, CloudBackupError> {
-        let existing_identities = crate::backup::import::collect_existing_wallet_identities()
+        let existing_identities = crate::wallet_identity::collect_existing_wallet_identities()
             .map_err_prefix("collect wallet identities", CloudBackupError::Internal)?;
         let mut restore_session = WalletRestoreSession::new(existing_identities);
         let mut merged_namespaces = Vec::with_capacity(namespaces.len());
