+++ rust/src/manager/cloud_backup_manager/ops/restore.rs
@@ -21,7 +21,7 @@ impl RustCloudBackupManager {
         namespaces: Vec<NamespaceMatch>,
     ) -> Result<CloudBackupRestoreReport, CloudBackupError> {
         let current_namespace = self.current_namespace_id()?;
-        let existing_identities = crate::backup::import::collect_existing_wallet_identities()
+        let existing_identities = crate::wallet_identity::collect_existing_wallet_identities()
             .map_err_prefix("collect wallet identities", CloudBackupError::Internal)?;
         let mut restore_session = WalletRestoreSession::new(existing_identities);
         let mut current_wallet_record_ids: HashSet<_> = current_namespace_wallet_record_ids(
